Make update with MVVM and state Management Provider
