+++ ios/Runner/Info.plist
@@ -4,6 +4,11 @@
 	<dict>
 		<key>CFBundleDevelopmentRegion</key>
 		<string>$(DEVELOPMENT_LANGUAGE)</string>
+		<key>LSApplicationQueriesSchemes</key>
+        <array>
+          <string>sms</string>
+          <string>tel</string>
+        </array>
 		<key>CFBundleDisplayName</key>
 		<string>News App Route</string>
 		<key>CFBundleExecutable</key>
