+++ lib/Shared/network/remote/api_consumer.dart
@@ -1,0 +1,6 @@
+abstract class ApiConsumer {
+  Future<dynamic> get(
+    Uri url, {
+    Map<String, String>? headers,
+  });
+}
